[1/n] Remove Sui direct dependency on `move-vm-types`
